Fixed errors

+++ src/main/java/school/mjc/stage0/loops/task2/PowerOfTwo.java
@@ -7,7 +7,7 @@ public class PowerOfTwo {
         } else {
             int i = 0;
             while (i <= power) {
-                System.out.println(Math.pow(2, i));
+                System.out.println((int) Math.pow(2, i));
                 i++;
             }
         }
